requirements.txt
kafka-python
sqlalchemy
requests
python-dotenv
pytest
